User authentication and bootstrap

+++ essen/menu/views.py
@@ -9,6 +9,8 @@ from recipes.models import Recipe, Ingredient
 from menu.models import Menu, Meal, LatePlate, AutoLatePlate
 from django.http import HttpResponse, HttpResponseRedirect
 from django.urls import reverse
+from django.contrib.auth.models import User
+from django.utils.encoding import python_2_unicode_compatible
 
 # Create your views here.
 
@@ -37,17 +39,17 @@ def index(request, date='None'):
 
     return render(request, template_name, {context_object_name: menu, 'target_date': target_date})
 
-
 def add_menu(request):
     template_name = "menu/add_menu.html"
     context_object_name = 'recipe_choices'
 
     recipe_choices = Recipe.objects.all()
 
-    return render(request, template_name, {context_object_name: recipe_choices})
+    return render(request, template_name, {context_object_name: recipe_choices, 'steward': check_if_steward(request.user)})
 
 
 def submit_menu(request):
+    print(request.POST)
     days_to_num = {"Sunday Brunch": 0, "Sunday Dinner": 0, "Monday Dinner": 1, "Tuesday Dinner": 2,
                    "Wednesday Dinner": 3, "Thursday Dinner": 4}
 
@@ -67,17 +69,19 @@ def submit_menu(request):
             day_num = key.split("_")[1]
             meal_key = "item_" + day_num
 
-            recipes = d[meal_key]
+            recipes = d[meal_key] #list of all recipe names corresponding to that meal in the menu
             meal = Meal(menu=menu,
                         date=start_date+timedelta(days_to_num[request.POST.get(key)]),
                         meal_name=request.POST.get(key))
             meal.save()
 
+            # add automatic lateplates
             for user in AutoLatePlate.objects.all():
-                if str(days_to_num[request.POST.get(key)]) in str(user.days):
+                if str(request.POST.get(key)) in str(user.days):
                     l = LatePlate(meal=meal, name=user.username)
                     l.save()
 
+            # add recipes
             for r in recipes:
                 recipe = Recipe.objects.filter(recipe_name=r).first()
                 meal.recipes.add(recipe)
@@ -85,47 +89,74 @@ def submit_menu(request):
     return HttpResponseRedirect(reverse('menu:index'))
 
 
-class DetailView(generic.DetailView):
-    model = Meal
+def view_meal(request, pk):
     template_name = 'menu/display_meal.html'
+    context_object_name = 'info'
 
-
-def add_lateplate(request, pk):
     meal = get_object_or_404(Meal, pk=pk)
 
-    l = LatePlate(name=request.POST.get("name"), meal=meal)
-    l.save()
+    info = {"meal" : meal, "users" : User.objects.all()}
 
-    return HttpResponseRedirect(reverse('menu:display_meal', args=[pk]))
+    return render(request, template_name, {context_object_name: info})
 
 
-def remove_lateplate(request, pk):
-    lateplate = get_object_or_404(LatePlate, pk=pk)
-    meal_id = lateplate.meal.id
-    lateplate.delete()
+def add_lateplate(request, meal_pk, user_pk):
+    meal = get_object_or_404(Meal, pk=meal_pk)
+    user = get_object_or_404(User, pk=user_pk)
+
+    if user.is_authenticated:
+        l = LatePlate(name=request.POST.get("name"), meal=meal)
+        l.save()
+
+    return HttpResponseRedirect(reverse('menu:display_meal', args=[meal_pk]))
+
+
+def remove_lateplate(request, lateplate_pk, user_pk):
+    lateplate = get_object_or_404(LatePlate, pk=lateplate_pk)
+
+    if request.user.is_authenticated:
+        meal_id = lateplate.meal.id
+        lateplate.delete()
 
     return HttpResponseRedirect(reverse('menu:display_meal', args=[meal_id]))
 
 
-class AutoLatePlates(generic.ListView):
+def auto_lateplates(request):
+    template_name = 'menu/auto_lateplates.html'
+    map = {"Sunday Brunch": 0, "Sunday Dinner": 1, "Monday Dinner": 2, "Tuesday Dinner": 3,
+                   "Wednesday Dinner": 4, "Thursday Dinner": 5}
+    requested_days = [{"day": "Sunday Brunch", "state": False}, {"day": "Sunday Dinner", "state": False}, {"day": "Monday Dinner", "state": False},
+                      {"day": "Tuesday Dinner", "state": False},  {"day": "Wednesday Dinner", "state": False}, {"day": "Thursday Dinner", "state": False},]
+
+    lateplate = AutoLatePlate.objects.filter(username=request.user.username).first()
+
+    if lateplate != None:
+        print(lateplate.days)
+        for day in lateplate.days.split(";"):
+            print("day, " ,day)
+            requested_days[map[day]] = {"day": day, "state": True}
+
+    return render(request, template_name, {"days" : requested_days})
+
     model = AutoLatePlate
     template_name = 'menu/auto_lateplates.html'
 
 
 def submit_auto_lateplates(request):
-    username = request.POST.get("username")
-    days = request.POST.get("days")
-
-    auto = AutoLatePlate(username=username, days=days)
-
-    auto.save()
-
-    return HttpResponseRedirect(reverse('menu:index'))
-
-
-def remove_auto_lateplates(request, pk):
-    auto = get_object_or_404(AutoLatePlate, pk=pk)
-    auto.delete()
+    d = dict(request.POST.iterlists())
+    # delete the previous lateplate registrys
+    AutoLatePlate.objects.filter(username=request.user.username).delete()
+    print(d)
+    # add the new one
+    if "date" in d.keys():
+        days = ""
+        for day in d.get("date"):
+            if days != "":
+                days += ";"
+            days += day
+            print(day)
+        auto = AutoLatePlate(username=request.user.username, days=days)
+        auto.save()
 
     return HttpResponseRedirect(reverse('menu:index'))
 
@@ -170,3 +201,5 @@ def ingredient_info(request, ing, menu):
     return render(request, template_name, {context_object_name: ingredient_uses})
 
 
+def check_if_steward(user):
+    return user.groups.all().filter(name="stewards").count() > 0
